requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.2.4
django-rest-auth==0.9.5
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
six==1.17.0
sqlparse==0.5.3
